place order functionality complete

+++ components/payment/PriceConfirm.jsx
@@ -5,9 +5,18 @@ import ProgressStepsContext from "../../context/progressStepsContext";
 import { useRouter } from "next/router";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteAllCartItems } from "../../redux/feature/cart/cartSlice";
+import { placeOrder, reset } from "../../redux/feature/order/orderSlice";
 
 const PriceConfirm = () => {
   const { message, isLoading, isSuccess } = useSelector((state) => state.cart);
+  const {
+    message: orderMessage,
+    isLoading: orderIsLoading,
+    isSuccess: orderIsSuccess,
+    isError: orderIsError,
+    order,
+  } = useSelector((state) => state.order);
+
   const { totalPrice, items } = useContext(CartContext);
   const {
     setProgress,
@@ -32,11 +41,24 @@ const PriceConfirm = () => {
     }));
   };
 
-  const placeOrder = () => {
-    toast.success("Order Placed");
-    dispatch(deleteAllCartItems());
-    backToCart();
-    router.push("/my-orders");
+  useEffect(() => {
+    if (orderIsError) {
+      toast.error(orderMessage);
+    }
+
+    if (orderIsSuccess && order) {
+      toast.success("Order Placed");
+      dispatch(deleteAllCartItems());
+
+      backToCart();
+      router.push("/my-orders");
+    }
+
+    dispatch(reset());
+  }, [orderIsSuccess, orderIsError]);
+
+  const handlePlaceOrder = () => {
+    dispatch(placeOrder());
   };
 
   return (
@@ -60,7 +82,7 @@ const PriceConfirm = () => {
       <div className="fixed bottom-0 z-20 bg-white w-[93%] flex justify-between items-center p-3">
         <h1 className="text-xl font-bold">₹{totalPrice}</h1>
         <button
-          onClick={summary.pending ? placeOrder : goToSummary}
+          onClick={summary.pending ? handlePlaceOrder : goToSummary}
           className="bg-[#f43397] text-white px-3 py-2 w-[48%] mb-1 rounded-md flex items-center justify-center"
         >
           {summary.pending ? "Place Order" : "Continue"}
